Finished API Authentication. Still needs Frontend Controller.

+++ BikeShopAnalyticsAPI/Controllers/BikeController.cs
@@ -15,57 +15,80 @@ namespace BikeShopAnalytics.Controllers
     public class BikeController : ControllerBase
     {
         private IRepository<Bike> _bikeRepo;
+        private IRepository<Auth> _authRepo;
 
-        public BikeController(IRepository<Bike> bikeRepo)
+        public BikeController(IRepository<Bike> bikeRepo, IRepository<Auth> authRepo)
         {
             _bikeRepo = bikeRepo;
+            _authRepo = authRepo;
         }
 
         [HttpGet("[action]/{bikeID}")]
-        public async Task<Bike> Read(int bikeID)
+        public async Task<ActionResult<Bike>> Read([FromHeader(Name = "Token")]string token, int bikeID)
         {
-            return await _bikeRepo.Read(b => b.BikeID == bikeID);
+            if(await _authRepo.Read(a => a.Token == token) != null)
+            {
+                return await _bikeRepo.Read(b => b.BikeID == bikeID);
+            }
+            return StatusCode(403);
         }
 
         [HttpPost("[action]")]
-        public async Task<IActionResult> Create(Bike bike)
+        public async Task<IActionResult> Create([FromHeader(Name = "Token")]string token, Bike bike)
         {
-            if (ModelState.IsValid)
+            if(await _authRepo.Read(a => a.Token == token) != null)
             {
-                await _bikeRepo.Create(bike);
-                return Ok("Success! Bike Created!");
+                if (ModelState.IsValid)
+                {
+                    await _bikeRepo.Create(bike);
+                    return Ok("Success! Bike Created!");
+                }
+                return Problem("Error! Could not create the bike..");
             }
-            return Problem("Error! Could not create the bike..");
+            return StatusCode(403);
         }
 
         [HttpPut("[action]")]
-        public async Task<IActionResult> Update(Bike bike)
+        public async Task<IActionResult> Update([FromHeader(Name = "Token")]string token, Bike bike)
         {
-            if (ModelState.IsValid)
+            if(await _authRepo.Read(a => a.Token == token) != null)
             {
-                await _bikeRepo.Update(bike);
-                return Ok("Success! Bike Updated!");
+                if (ModelState.IsValid)
+                {
+                    await _bikeRepo.Update(bike);
+                    return Ok("Success! Bike Updated!");
+                }
+                return Problem("Error! Could not update the Bike..");
             }
-            return Problem("Error! Could not update the Bike..");
+            return StatusCode(403);
         }
 
         [HttpDelete("[action]/{bikeID}")]
-        public async Task<IActionResult> Delete(int bikeID)
+        public async Task<IActionResult> Delete([FromHeader(Name = "Token")]string token, int bikeID)
         {
-            var bike = new Bike();//Read(bikeID);
-            if(!(bike is null))
+            if(await _authRepo.Read(a => a.Token == token) != null)
             {
-                await _bikeRepo.Delete(bike);
-                return Ok("Success! Bike Deleted!");
+                var bike = await _bikeRepo.Read(a => a.BikeID == bikeID);
+
+                if(!(bike is null))
+                {
+                    await _bikeRepo.Delete(bike);
+                    return Ok("Success! Bike Deleted!");
+                }
+                return Problem("Error! Could not delete the Bike..");
             }
-            return Problem("Error! Could not delete the Bike..");
+            return StatusCode(403);
         }
 
         [HttpGet("[action]")]
-        public async Task<List<Bike>> ReadAll()
+        public async Task<ActionResult<List<Bike>>> ReadAll([FromHeader(Name = "Token")]string token)
         {
-            var bikeList = await _bikeRepo.ReadAll();
-            return bikeList.ToList();;
+            if(await _authRepo.Read(a => a.Token == token) != null)
+            {
+                var bikeList = await _bikeRepo.ReadAll();
+                return bikeList.ToList();;
+            }
+            return StatusCode(403);
         }
     }
 }
